video/requirements_video.txt
# Requirements for Gemini Video Generation with Veo 3.0
google-genai>=0.8.0
python-dotenv>=1.0.0